Refactor view names and enclosing packages
